requirements.txt
pypdf==6.1.1
openai==2.1.0
python-dotenv==1.1.1
pymongo==4.15.2